# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: BASELINE Field Operations Intelligence API
  version: 1.0.0
  description: |-
    BASELINE captures operational truth through field visits and derives intelligence from that truth.
    Every endpoint documents its required role, request/response shapes, and applicable audit log action.

    **Role hierarchy**: Admin > Lead > UL (Unit Leader / field worker)
    **Auth**: Bearer token (Azure Entra JWT) on all non-public endpoints.
    **Labor safety**: No endpoint exposes per-worker performance metrics or identifiers
    to Lead/UL roles. `captured_by_oid` is encrypted and Admin-access-only.
  contact:
    name: BASELINE Engineering
servers:
  - url: http://localhost:3001/api
    description: Local development
  - url: /api
    description: Production (relative)
...
components:
  securitySchemes:
    AzureAD:
      type: oauth2
      description: 'Azure Active Directory (Entra) OAuth2. Pass the resulting access token as `Authorization: Bearer <token>`.'
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/{tenantId}/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/{tenantId}/oauth2/v2.0/token
          scopes:
            openid: OpenID Connect
            profile: User profile
            email: Email address
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Resource not found
  responses:
    BadRequest:
      description: Request validation failed — missing or invalid field
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: pool_id is required
    Unauthorized:
      description: Authentication required or Bearer token is invalid/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Unauthorized
    Forbidden:
      description: Caller lacks the required role for this endpoint
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Forbidden
...
    PayloadTooLarge:
      description: File exceeds the maximum allowed upload size (25 MB)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: File exceeds 25 MB limit
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error
paths:
...
  /ul/inbox:
    get:
      summary: UL inbox placeholder
      description: Returns a confirmation that the caller has the UL role. Reserved for future inbox features.
      tags:
        - UL
      security:
        - AzureAD: []
      x-required-roles:
        - UL
      responses:
        '200':
          description: Caller is a UL
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  scope:
                    type: string
              example:
                ok: true
                scope: UL
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /ul/todays-run:
    get:
      summary: Get today's active route run for the authenticated UL
      description: >
        Looks up the most recent planned or in_progress route run assigned to the caller's Azure Entra OID. Returns the
        full route run with stop details. Also accessible to Lead and Admin for supervisory views.
      tags:
        - UL
      security:
        - AzureAD: []
      x-required-roles:
        - UL
        - Lead
        - Admin
      responses:
        '200':
          description: Active route run found
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  route_run:
                    type: object
                    description: Full route run with stops
              example:
                ok: true
                route_run:
                  id: 42
                  status: in_progress
                  route_pool_id: POOL-001
                  stops: []
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No active assigned route run found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: No active assigned route run found
        '500':
          $ref: '#/components/responses/InternalError'
  /route-runs/{runId}/stops/{stopId}/photos:
    post:
      summary: Upload photos for a route run stop
      description: >
        Accepts up to 10 image files via multipart/form-data. Each file is validated for MIME type (JPEG, PNG, WebP,
        HEIC) and size (≤25 MB). Files are stored in S3 and references saved to the database.
      tags:
        - UL
      security:
        - AzureAD: []
      x-required-roles:
        - UL
        - Lead
        - Admin
      x-audit-action: upload.rejected
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: integer
          description: Route run ID
          example: 42
        - in: path
          name: stopId
          required: true
          schema:
            type: integer
          description: Route run stop ID
          example: 7
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photos:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Up to 10 image files
                kind:
                  type: string
                  enum:
                    - safety
                    - cleaning
                    - infrastructure
                    - completion
                    - skip_after
                  default: completion
                  description: Photo category
      responses:
        '200':
          description: Photos uploaded and saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  photos:
                    type: array
                    items:
                      type: object
              example:
                ok: true
                photos:
                  - id: 1
                    s3_key: runs/42/stops/7/uuid.jpg
                    kind: completion
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      summary: List photos for a route run stop
      description: Returns all photos attached to the specified route run stop.
      tags:
        - UL
      security:
        - AzureAD: []
      x-required-roles:
        - UL
        - Lead
        - Admin
      parameters:
        - in: path
          name: runId
          required: true
          schema:
            type: integer
          description: Route run ID
          example: 42
        - in: path
          name: stopId
          required: true
          schema:
            type: integer
          description: Route run stop ID
          example: 7
      responses:
        '200':
          description: List of photos
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  photos:
                    type: array
                    items:
                      type: object
              example:
                ok: true
                photos:
                  - id: 1
                    s3_key: runs/42/stops/7/uuid.jpg
                    kind: completion
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
...